[elastic] Guess the module path for 'go mod init' call (#87)
The original implementation is passing the current folder as the module
path, this will disable the cross-package jump within the same repo.

There are two two steps to guess the module path. The first step is
guessing by `findModulePath()` which is copied from the core logic of
'go mod init', see
https://github.com/golang/go/blob/release-branch.go1.12/src/cmd/go/alldocs.go#L1040

The second step is constructed by our own, construct the module path
from the local filesystem. The construction follows the
'codehost/owner/repo' pattern.